requirements.txt
Django==5.2.6
psycopg[binary]>=3.2,<3.4
python-dotenv>=1.0
djangorestframework
PyJWT
django-cors-headers
drf-spectacular>=0.27